+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptExpression.java
@@ -6,6 +6,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.starrocks.sql.optimizer.base.ColumnRefSet;
 import com.starrocks.sql.optimizer.base.LogicalProperty;
+import com.starrocks.sql.optimizer.base.PhysicalPropertySet;
 import com.starrocks.sql.optimizer.operator.Operator;
 import com.starrocks.sql.optimizer.statistics.Statistics;
 
@@ -33,6 +34,8 @@ public class OptExpression {
     // For easily convert a GroupExpression to OptExpression when pattern match
     // we just use OptExpression to wrap GroupExpression
     private GroupExpression groupExpression;
+    // required properties for children.
+    private List<PhysicalPropertySet> requiredProperties;
 
     public OptExpression() {
         this.inputs = Lists.newArrayList();
@@ -106,6 +109,14 @@ public class OptExpression {
     public ColumnRefSet getOutputColumns() {
         Preconditions.checkState(property != null);
         return property.getOutputColumns();
+    }
+
+    public void setRequiredProperties(List<PhysicalPropertySet> requiredProperties) {
+        this.requiredProperties = requiredProperties;
+    }
+
+    public List<PhysicalPropertySet> getRequiredProperties() {
+        return this.requiredProperties;
     }
 
     // This function assume the child expr logical property has been derived
